Refactor: Update ViewModel logic, UI interactions, and dependencies
This commit introduces several changes primarily focused on refactoring ViewModels, enhancing UI interactions, and updating dependencies.

Key Changes:

*   **ViewModel and Data Handling (`MainViewModel.kt`, `EvidenceViewModel.kt`):**
    *   Corrected `Evidence` model import path from `com.hereliesaz.lexorcist.model.Evidence` to `com.hereliesaz.lexorcist.data.Evidence` in various ViewModels and test files.
    *   Changed `TAG` property to `tag` in `MainViewModel.kt`.
    *   Updated `getEvidenceForCase` in `MainViewModel.kt` and `EvidenceRepositoryImpl.kt` to correctly handle `caseId` type conversion (Int to Long).
    *   Refined file parsing logic (`parseTextFile`, `parsePdfFile`, `parseSpreadsheetFile`, `parseDocFile`) in `MainViewModel.kt` to be non-suspend and assign appropriate categories and placeholder IDs.
    *   Enhanced SMS import to include message date and assign "SMS" category in `MainViewModel.kt`.
    *   Improved error handling by setting `_errorMessage.value` in various parsing and upload functions in `MainViewModel.kt`.
    *   Adjusted `addTextEvidenceToSelectedCase` and `addDriveFileEvidenceToSelectedCase` in `MainViewModel.kt` to create `Evidence` objects with more specific categories and to handle scenarios where a case or API service is not available.
    *   Corrected `storeTaggedDataToSheet` in `MainViewModel.kt` to use the `tag` property for logging.
    *   Removed unused imports like `android.provider.Settings.Global.putString`, OpenCV, and `androidx.room.util.copy`.
    *   Modified `sharedPref.edit()` calls to use the lambda with receiver `sharedPref.edit { ... }`.
    *   Updated `confirmImageReview` to set an appropriate category and tags for OCR'd images.
    *   Ensured PDF generation logic in `generatePdfFromHtmlString` in `MainViewModel.kt` uses `Dispatchers.Main` for WebView operations.
    *   Updated `importSpreadsheet` to include try-catch for schema loading and parsing.
*   **UI Updates (`TimelineScreen.kt`, `MainScreen.kt`, `ScriptEditorActivity.kt`, `Theme.kt`, `MainActivity.kt`):**
    *   In `TimelineScreen.kt`:
        *   Changed `EventPointType` from `Filled` to `POINT` and removed `pointIcon`.
        *   Handled nullable `documentDate` and `sourceDocument` for `Evidence` items.
        *   Updated `getIconForCategory` to include more categories and use `Icons.AutoMirrored.Filled` for relevant icons.
        *   Corrected a typo in `EvidenceDetailsDialog` for displaying tags (`evidence.tags.joinToString().ifEmpty { "N/A" }`).
        *   Added `@OptIn(ExperimentalComposeApi::class)`.
    *   In `MainScreen.kt`:
        *   Passed `viewModel()` (implicitly `MainViewModel`) to `TimelineScreen` instead of `viewModel` from `MainScreen`'s parameters.
        *   Added missing imports for new ViewModels (`AuthViewModel`, `CaseViewModel`, `EvidenceViewModel`, `OcrViewModel`).
    *   Added `@OptIn(ExperimentalMaterial3Api::class)` to `ScriptEditorActivity.kt`, `Theme.kt`, and `MainActivity.kt`.
    *   In `MainActivity.kt`:
        *   `MainScreen` composable call was updated to no longer pass `navController` as it was not used by `MainScreen`.
        *   Corrected the instantiation of `GetContentWithMultiFilter` by passing `mimeTypes` in its constructor.
        *   Corrected the `launch` calls for `selectDocumentLauncher` and `selectSpreadsheetLauncher` to pass a single primary MIME type string (e.g., `"*/*"`), as the contract handles the multiple types.
        *   Refactored `signIn` method to improve logging and error handling for One Tap UI.
*   **File Handling (`GetContentWithMultiFilter.kt`):**
    *   Modified `GetContentWithMultiFilter` to accept `mimeTypes` in its constructor and use them when creating the intent, while still accepting the single `input` MIME type for the superclass method.
*   **Dependency Updates (`app/build.gradle.kts`):**
    *   Added `androidx.room:room-ktx:2.7.2`.
*   **Test Updates (`GoogleApiServiceTest.kt`, `EvidenceViewModelTest.kt`):**
    *   Updated import paths for `Evidence` model.

+++ app/src/main/java/com/hereliesaz/lexorcist/data/EvidenceRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package com.hereliesaz.lexorcist.data
 
 import com.hereliesaz.lexorcist.GoogleApiService
-import com.hereliesaz.lexorcist.data.Evidence // Corrected import
 import kotlinx.coroutines.flow.Flow
 import java.lang.Exception
 
@@ -17,7 +16,7 @@ class EvidenceRepositoryImpl(
     override fun getEvidenceForCase(caseId: Long): Flow<List<Evidence>> {
         // Assuming EvidenceDao.getEvidenceForCase expects a Long and handles mapping to Evidence.caseId (Int) if necessary internally or via query conversion.
         // Or, that the Case ID in the database is indeed stored as Long and Evidence.caseId (Int) is for the model object after retrieval.
-        return evidenceDao.getEvidenceForCase(caseId)
+        return evidenceDao.getEvidenceForCase(caseId.toLong())
     }
 
     override suspend fun refreshEvidenceForCase(spreadsheetId: String, caseId: Int) {
